pyproject.toml
[project]
name = "bonfire"
version = "0.1.0"
description = ""
authors = [
    {name = "skribblez"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
